life-framework-v1/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "life-framework"
version = "0.1.0"
description = "LIFE — LoRA Influencer Fabrication Engine"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "httpx>=0.25.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "python-dotenv>=1.0.0",
    "click>=8.0",
]

[project.scripts]
life = "life.cli:main"

[tool.setuptools.packages.find]
include = ["life*"]
